requirements.txt
feedparser
pyyaml
pymisp